rename: rename background images
